Rename record-error-archival lambda to add-archival-events
