+++ src/add-archival-events/Makefile
@@ -1,0 +1,2 @@
+build:
+	make -C ../../.. add-archival-events
