--- src/record-error-archival/Makefile
@@ -1,2 +1,0 @@
-build:
-	make -C ../../.. record-error-archival
